+++ auction-service/src/main/java/com/farmted/auctionservice/dto/ResponseBiddingDto/BiddingResponseDto.java
@@ -5,11 +5,12 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter @AllArgsConstructor @NoArgsConstructor
 public class BiddingResponseDto {
-    private Integer biddingPrice; // 일반 입찰 금액
+    private BigDecimal biddingPrice; // 일반 입찰 금액
     private LocalDateTime biddingTime; // 입찰 시간
     private String memberUuid; // 응찰자
