+++ auction-service/src/main/java/com/farmted/auctionservice/dto/responseAuctionDto/AuctionGetResponseDto.java
@@ -7,11 +7,12 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter @AllArgsConstructor @NoArgsConstructor @Builder
 public class AuctionGetResponseDto {
-    private Integer auctionPrice; // 경매 가격
+    private BigDecimal auctionPrice; // 경매 가격
     private LocalDate auctionDeadline; // 경매 종료
     private String productUuid;
     private boolean auctionStatus;
