+++ auction-service/src/main/java/com/farmted/auctionservice/dto/responseAuctionDto/AuctionStatusResponseDto.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 @Getter
 @AllArgsConstructor
@@ -13,7 +14,7 @@ public class AuctionStatusResponseDto {
     private String productUuid;
     private LocalDate auctionDeadline; // 경매 종료 시간
     private Boolean auctionStatus; // 0: 경매중, 1:  경매 종료
-    private Integer auctionPrice; // 가격
+    private BigDecimal auctionPrice; // 가격
     private String auctionBuyer;  // 낙찰자
